Move role views to their own file
