+++ corehq/apps/users/views/__init__.py
@@ -10,12 +10,10 @@ import six.moves.urllib.request
 from couchdbkit.exceptions import ResourceNotFound
 from crispy_forms.utils import render_crispy_form
 
-from corehq.apps.cloudcare.dbaccessors import get_cloudcare_apps, get_application_access_for_domain
 from corehq.apps.custom_data_fields.edit_entity import CustomDataEditor
-from corehq.apps.custom_data_fields.models import CustomDataFieldsProfile, CustomDataFieldsDefinition, PROFILE_SLUG
+from corehq.apps.custom_data_fields.models import CustomDataFieldsProfile, PROFILE_SLUG
 from corehq.apps.programs.models import Program
-from corehq.apps.registry.utils import get_data_registry_dropdown_options
-from corehq.apps.reports.models import TableauVisualization, TableauUser
+from corehq.apps.reports.models import TableauUser
 from corehq.apps.sso.models import IdentityProvider
 from corehq.apps.sso.utils.user_helpers import get_email_domain_from_username
 from corehq.toggles import TABLEAU_USER_SYNCING
@@ -34,10 +32,8 @@ from django.shortcuts import render
 from django.urls import reverse
 from django.utils.decorators import method_decorator
 from django.utils.functional import cached_property
-from django.utils.safestring import mark_safe
-from django.utils.translation import gettext as _, ngettext, gettext_lazy, gettext_noop
+from django.utils.translation import gettext as _, gettext_lazy, gettext_noop
 
-from corehq.apps.users.analytics import get_role_user_count
 from soil.exceptions import TaskFailedError
 from soil.util import expose_cached_download, get_download_context
 from django.views.decorators.csrf import csrf_exempt
@@ -57,7 +53,6 @@ from corehq.apps.analytics.tasks import (
 )
 from corehq.apps.app_manager.dbaccessors import get_app_languages
 from corehq.apps.domain.decorators import (
-    domain_admin_required,
     login_and_domain_required,
     require_superuser,
 )
@@ -75,7 +70,6 @@ from corehq.apps.registration.forms import (
     AdminInvitesUserForm,
 )
 from corehq.apps.reports.exceptions import TableauAPIError
-from corehq.apps.reports.util import get_possible_reports
 from corehq.apps.sms.mixin import BadSMSConfigException
 from corehq.apps.sms.verify import (
     VERIFICATION__ALREADY_IN_USE,
@@ -85,16 +79,14 @@ from corehq.apps.sms.verify import (
     initiate_sms_verification_workflow,
 )
 from corehq.apps.translations.models import SMSTranslations
-from corehq.apps.userreports.util import has_report_builder_access
 from corehq.apps.users.audit.change_messages import UserChangeMessage
 from corehq.apps.users.decorators import (
     can_use_filtered_user_download,
     require_can_edit_or_view_web_users,
     require_can_edit_web_users,
-    require_can_view_roles,
     require_permission_to_edit_user,
 )
-from corehq.apps.users.exceptions import MissingRoleException, InvalidRequestException
+from corehq.apps.users.exceptions import MissingRoleException
 from corehq.apps.users.forms import (
     BaseUserInfoForm,
     CommtrackUserForm,
@@ -102,7 +94,6 @@ from corehq.apps.users.forms import (
     TableauUserForm,
     WebUserFormSet,
 )
-from corehq.apps.users.landing_pages import get_allowed_landing_pages, validate_landing_page
 from corehq.apps.users.models import (
     CommCareUser,
     CouchUser,
@@ -112,7 +103,6 @@ from corehq.apps.users.models import (
     Invitation,
     StaticRole,
     WebUser,
-    HqPermissions,
     UserRole,
 )
 from corehq.apps.users.model_log import InviteModelAction
@@ -129,7 +119,6 @@ from corehq.const import USER_CHANGE_VIA_WEB, INVITATION_CHANGE_VIA_WEB
 from corehq.pillows.utils import WEB_USER_TYPE
 from corehq.toggles import PARALLEL_USER_IMPORTS
 from corehq.util.couch import get_document_or_404
-from corehq.util.view_utils import json_error
 from corehq.util.workbook_json.excel import (
     WorkbookJSONError,
     WorksheetNotFound,
@@ -652,147 +641,6 @@ class DownloadWebUsersStatusView(BaseUserSettingsView):
         return reverse(self.urlname, args=self.args, kwargs=self.kwargs)
 
 
-class ListRolesView(BaseRoleAccessView):
-    template_name = 'users/roles_and_permissions.html'
-    page_title = gettext_lazy("Roles & Permissions")
-    urlname = 'roles_and_permissions'
-
-    @method_decorator(require_can_view_roles)
-    def dispatch(self, request, *args, **kwargs):
-        return super(ListRolesView, self).dispatch(request, *args, **kwargs)
-
-    @property
-    def can_edit_roles(self):
-        return (has_privilege(self.request, privileges.ROLE_BASED_ACCESS)
-                and self.couch_user.is_domain_admin)
-
-    @property
-    def landing_page_choices(self):
-        return [
-            {'id': None, 'name': _('Use Default')}
-        ] + [
-            {'id': page.id, 'name': _(page.name)}
-            for page in get_allowed_landing_pages(self.domain)
-        ]
-
-    @property
-    @memoized
-    def non_admin_roles(self):
-        return list(sorted(
-            [role for role in UserRole.objects.get_by_domain(self.domain) if not role.is_commcare_user_default],
-            key=lambda role: role.name if role.name else '\uFFFF'
-        )) + [UserRole.commcare_user_default(self.domain)]  # mobile worker default listed last
-
-    def can_edit_linked_roles(self):
-        return self.request.couch_user.can_edit_linked_data(self.domain)
-
-    def get_roles_for_display(self):
-        show_es_issue = False
-        role_view_data = [StaticRole.domain_admin(self.domain).to_json()]
-        for role in self.non_admin_roles:
-            role_data = role.to_json()
-            role_view_data.append(role_data)
-
-            if role.is_commcare_user_default:
-                role_data["preventRoleDelete"] = True
-            else:
-                try:
-                    user_count = get_role_user_count(role.domain, role.couch_id)
-                    role_data["preventRoleDelete"] = bool(user_count)
-                except TypeError:
-                    # when query_result['hits'] returns None due to an ES issue
-                    show_es_issue = True
-
-            role_data["has_unpermitted_location_restriction"] = (
-                not self.can_restrict_access_by_location
-                and not role.permissions.access_all_locations
-            )
-
-        if show_es_issue:
-            messages.error(
-                self.request,
-                mark_safe(_(  # nosec: no user input
-                    "We might be experiencing issues fetching the entire list "
-                    "of user roles right now. This issue is likely temporary and "
-                    "nothing to worry about, but if you keep seeing this for "
-                    "more than a day, please <a href='#modalReportIssue' "
-                    "data-toggle='modal'>Report an Issue</a>."
-                ))
-            )
-        return role_view_data
-
-    def get_possible_profiles(self):
-        from corehq.apps.users.views.mobile.custom_data_fields import (
-            CUSTOM_USER_DATA_FIELD_TYPE,
-        )
-        definition = CustomDataFieldsDefinition.get(self.domain, CUSTOM_USER_DATA_FIELD_TYPE)
-        if definition is not None:
-            return [{
-                    'id': profile.id,
-                    'name': profile.name,
-                    }
-                for profile in definition.get_profiles()]
-        else:
-            return []
-
-    @property
-    def page_context(self):
-        from corehq.apps.linked_domain.dbaccessors import is_active_downstream_domain
-        if (not self.can_restrict_access_by_location
-                and any(not role.permissions.access_all_locations
-                        for role in self.non_admin_roles)):
-            messages.warning(self.request, _(
-                "This project has user roles that restrict data access by "
-                "organization, but the software plan no longer supports that. "
-                "Any users assigned to roles that are restricted in data access "
-                "by organization can no longer access this project.  Please "
-                "update the existing roles."))
-
-        tableau_list = []
-        if toggles.EMBEDDED_TABLEAU.enabled(self.domain):
-            tableau_list = [{
-                'id': viz.id,
-                'name': viz.name,
-            } for viz in TableauVisualization.objects.filter(domain=self.domain)]
-
-        return {
-            'is_managed_by_upstream_domain': is_active_downstream_domain(self.domain),
-            'can_edit_linked_data': self.can_edit_linked_roles(),
-            'user_roles': self.get_roles_for_display(),
-            'non_admin_roles': self.non_admin_roles,
-            'can_edit_roles': self.can_edit_roles,
-            'default_role': StaticRole.domain_default(self.domain),
-            'tableau_list': tableau_list,
-            'report_list': get_possible_reports(self.domain),
-            'profile_list': self.get_possible_profiles(),
-            'is_domain_admin': self.couch_user.is_domain_admin,
-            'domain_object': self.domain_object,
-            'uses_locations': self.domain_object.uses_locations,
-            'can_restrict_access_by_location': self.can_restrict_access_by_location,
-            'landing_page_choices': self.landing_page_choices,
-            'show_integration': (
-                toggles.OPENMRS_INTEGRATION.enabled(self.domain)
-                or toggles.DHIS2_INTEGRATION.enabled(self.domain)
-                or toggles.GENERIC_INBOUND_API.enabled(self.domain)
-            ),
-            'web_apps_choices': get_cloudcare_apps(self.domain),
-            'attendance_tracking_privilege': (
-                toggles.ATTENDANCE_TRACKING.enabled(self.domain)
-                and domain_has_privilege(self.domain, privileges.ATTENDANCE_TRACKING)
-            ),
-            'has_report_builder_access': has_report_builder_access(self.request),
-            'data_file_download_enabled':
-                domain_has_privilege(self.domain, privileges.DATA_FILE_DOWNLOAD),
-            'export_ownership_enabled': domain_has_privilege(self.domain, privileges.EXPORT_OWNERSHIP),
-            'data_registry_choices': get_data_registry_dropdown_options(self.domain),
-            'commcare_analytics_roles': _commcare_analytics_roles_options(),
-            'has_restricted_application_access': (
-                get_application_access_for_domain(self.domain).restrict
-                and toggles.WEB_APPS_PERMISSIONS_VIA_GROUPS.enabled(self.domain)
-            ),
-        }
-
-
 def _commcare_analytics_roles_options():
     return [
         {
@@ -1022,122 +870,6 @@ def reactivate_web_user(request, domain, couch_user_id):
         user.reactivate(domain, changed_by=request.couch_user)
         messages.success(request, 'You have successfully reactivated {username}.'.format(username=user.username))
     return HttpResponseRedirect(reverse(ListWebUsersView.urlname, args=[domain]))
-
-
-# If any permission less than domain admin were allowed here, having that
-# permission would give you the permission to change the permissions of your
-# own role such that you could do anything, and would thus be equivalent to
-# having domain admin permissions.
-@json_error
-@domain_admin_required
-@require_POST
-def post_user_role(request, domain):
-    if not domain_has_privilege(domain, privileges.ROLE_BASED_ACCESS):
-        return JsonResponse({})
-    role_data = json.loads(request.body.decode('utf-8'))
-
-    try:
-        role = _update_role_from_view(domain, role_data)
-    except ValueError as e:
-        return JsonResponse({
-            "message": str(e)
-        }, status=400)
-
-    response_data = role.to_json()
-    if role.is_commcare_user_default:
-        response_data["preventRoleDelete"] = True
-    else:
-        user_count = get_role_user_count(domain, role.couch_id)
-        response_data['preventRoleDelete'] = user_count > 0
-    return JsonResponse(response_data)
-
-
-def _update_role_from_view(domain, role_data):
-    landing_page = role_data["default_landing_page"]
-    if landing_page:
-        validate_landing_page(domain, landing_page)
-
-    if (
-        not domain_has_privilege(domain, privileges.RESTRICT_ACCESS_BY_LOCATION)
-        and not role_data['permissions']['access_all_locations']
-    ):
-        # This shouldn't be possible through the UI, but as a safeguard...
-        role_data['permissions']['access_all_locations'] = True
-
-    if "_id" in role_data:
-        try:
-            role = UserRole.objects.by_couch_id(role_data["_id"])
-        except UserRole.DoesNotExist:
-            role = UserRole()
-        else:
-            if role.domain != domain:
-                raise Http404()
-    else:
-        role = UserRole()
-
-    name = role_data["name"]
-    if not role.id:
-        if name.lower() == 'admin' or UserRole.objects.filter(domain=domain, name__iexact=name).exists():
-            raise ValueError(_("A role with the same name already exists"))
-
-    role.domain = domain
-    role.name = name
-    role.default_landing_page = landing_page
-    role.is_non_admin_editable = role_data["is_non_admin_editable"]
-    role.save()
-
-    permissions = HqPermissions.wrap(role_data["permissions"])
-    permissions.normalize(previous=role.permissions)
-    role.set_permissions(permissions.to_list())
-
-    assignable_by = role_data["assignable_by"]
-    role.set_assignable_by_couch(assignable_by)
-    return role
-
-
-@domain_admin_required
-@require_POST
-def delete_user_role(request, domain):
-    if not domain_has_privilege(domain, privileges.ROLE_BASED_ACCESS):
-        return JsonResponse({})
-    role_data = json.loads(request.body.decode('utf-8'))
-
-    try:
-        response_data = _delete_user_role(domain, role_data)
-    except InvalidRequestException as e:
-        return JsonResponse({"message": str(e)}, status=400)
-
-    return JsonResponse(response_data)
-
-
-def _delete_user_role(domain, role_data):
-    try:
-        role = UserRole.objects.by_couch_id(role_data["_id"], domain=domain)
-    except UserRole.DoesNotExist:
-        raise Http404
-
-    if role.is_commcare_user_default:
-        raise InvalidRequestException(_(
-            "Unable to delete role '{role}'. "
-            "This role is the default role for Mobile Users and can not be deleted.",
-        ).format(role=role_data["name"]))
-
-    user_count = get_role_user_count(domain, role_data["_id"])
-    if user_count:
-        raise InvalidRequestException(ngettext(
-            "Unable to delete role '{role}'. "
-            "It has one user and/or invitation still assigned to it. "
-            "Remove all users assigned to the role before deleting it.",
-            "Unable to delete role '{role}'. "
-            "It has {user_count} users and/or invitations still assigned to it. "
-            "Remove all users assigned to the role before deleting it.",
-            user_count,
-        ).format(role=role_data["name"], user_count=user_count))
-
-    copy_id = role.couch_id
-    role.delete()
-    # return removed id in order to remove it from UI
-    return {"_id": copy_id}
 
 
 @always_allow_project_access
